Android

+++ Android/app/src/main/java/com/example/myhomeapplication/Remote/MeasurementAPI.java
@@ -13,7 +13,7 @@ import retrofit2.http.Path;
 public interface MeasurementAPI {
 
     @GET("api/devices/{id}/measurements/last-{type}")
-    Call<Measurement> getLatestMeasurement(@Path("id") int deviceID, @Path("type") String measurementType);
+    Call<Measurement> getLatestMeasurement(@Path("id") long deviceID, @Path("type") String measurementType);
 
     @GET("api/devices/{id}/measurements/{type}")
     Call<List<Measurement>> getMeasurements(@Path("id") int deviceID, @Path("type") String measurementType);
